fix: payment DB now supports user_id BIGINT

+++ services/payment/db_setup.sql
@@ -6,7 +6,7 @@ CREATE TABLE IF NOT EXISTS receipts (
     id serial PRIMARY KEY,
     idempotency_key VARCHAR(255) UNIQUE NOT NULL,
     request_hash TEXT NOT NULL,
-    user_id INT NOT NULL,
+    user_id BIGINT NOT NULL,
     user_email VARCHAR(100) NOT NULL,
     amount NUMERIC(12,2) NOT NULL CHECK (amount >= 0),
     payment_description VARCHAR(255),
